Previously in the insert path, the rows to insert were iterated over once to determine the granule that each row belonged to and a second time to perform the insert. This commit improves performance by reading all the rows to insert once.

```
name            old time/op    new time/op    delta
InsertSimple-8    17.0ms ± 2%    15.1ms ± 2%  -11.15%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
InsertSimple-8    16.6MB ± 0%    13.5MB ± 1%  -18.51%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
InsertSimple-8      156k ± 0%      151k ± 0%   -3.23%  (p=0.000 n=10+10)
```